Update das classes de network e criação de classes de db

+++ app/src/main/java/com/example/braguia/model/Edge.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Edge(
 
     @SerializedName("id")
-    val id:Int,
+    val id:Long,
     @SerializedName("edge_transport")
     val edgeTransport:String,
     @SerializedName("edge_duration")
@@ -13,8 +13,7 @@ data class Edge(
     @SerializedName("edge_desc")
     val edgeDesc:String,
     @SerializedName("edge_trail")
-    val edgeTrail:Int,
-
+    val edgeTrail:Long,
     @SerializedName("edge_start")
     val edgeStart: Pin,
     @SerializedName("edge_end")
